feat: Discord로 알림이 가도록 완전히 구현
