+++ app/src/main/kotlin/com/wespot/config/OpenFeignConfig.kt
@@ -9,16 +9,17 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.http.converter.HttpMessageConverter
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
 import org.springframework.web.client.RestTemplate
+import java.util.logging.Logger
 
 
 @Configuration
 @EnableFeignClients(basePackages = ["com.wespot"])
 class OpenFeignConfig {
 
-    @Bean
-    fun errorDecoder(): ErrorDecoder {
-        return CustomErrorDecoder()
-    }
+//    @Bean
+//    fun errorDecoder(): ErrorDecoder {
+//        return CustomErrorDecoder()
+//    }
 
     @Bean
     fun feignMessageConverter(): MappingJackson2HttpMessageConverter {
