+++ infrastructure/discord/src/main/kotlin/com/wespot/discord/DiscordErrorNotificationClient.kt
@@ -1,0 +1,16 @@
+package com.wespot.discord
+
+import org.springframework.cloud.openfeign.FeignClient
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+
+@FeignClient(
+    name = "discord-error-notification-client",
+    url = "\${discord.error.webhook-url}",
+)
+interface DiscordErrorNotificationClient {
+
+    @PostMapping(consumes = ["application/json"])
+    fun notifyError(@RequestBody message: DiscordMessage)
+
+}
