+++ infrastructure/discord/src/main/kotlin/com/wespot/discord/DiscordWarningNotificationClient.kt
@@ -1,0 +1,16 @@
+package com.wespot.discord
+
+import org.springframework.cloud.openfeign.FeignClient
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+
+@FeignClient(
+    name = "discord-warning-notification-client",
+    url = "\${discord.warning.webhook-url}",
+)
+interface DiscordWarningNotificationClient {
+
+    @PostMapping(consumes = ["application/json"])
+    fun notifyWarning(@RequestBody message: DiscordMessage)
+
+}
