+++ infrastructure/redis/src/main/kotlin/com/wespot/config/RedisConfig.kt
@@ -11,15 +11,16 @@ import org.springframework.data.redis.serializer.StringRedisSerializer
 @Configuration
 class RedisConfig {
 
-    @Value("\${spring.data.redis.host}")
-    private lateinit var host: String
-
-    @Value("\${spring.data.redis.port}")
-    private var port: Int = 0
+//    @Value("\${spring.data.redis.host}")
+//    private lateinit var host: String
+//
+//    @Value("\${spring.data.redis.port}")
+//    private var port: Int = 0
 
     @Bean
     fun redisConnectionFactory(): LettuceConnectionFactory {
-        return LettuceConnectionFactory(host, port)
+//        return LettuceConnectionFactory(host, port)
+        return LettuceConnectionFactory()
     }
 
     @Bean
